Add files via upload
The fileorgv1 and fileorgv2 is what I use at work, v1 is to organize the pdf files in Downloads folder and v2 is to organize the snippet images in Pictures Folder. Note: the fileorg is okay but the problem with that is that it creates the folder within the folder because it gets confused with the folder we created and can't identify that its a directory folder.